requirements.txt
MetaTrader5>=5.0.45
pandas>=2.0.0
numpy>=1.26.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
openai>=1.30.0
anthropic>=0.28.0
google-generativeai>=0.7.0
aiosqlite>=0.20.0
ta>=0.11.0
